add views folder && add urls file in hyperion
